requirements.txt
absl-py==2.2.2
altair==5.5.0
altgraph==0.17.4
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
narwhals==1.38.2
networkx==3.4.2
numpy==2.1.3
opencv-python==4.11.0.86
opennsfw2==0.14.0
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pgi==0.0.11.2
pillow==11.2.1
protobuf==5.29.4
proxy_tools==0.1.0
pyarrow==20.0.0
pycairo==1.28.0
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.0
PyQt6_sip==13.10.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywebview==5.4
QtPy==2.4.3
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.4.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.45.0
streamlit-desktop-app==0.3.4
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tifffile==2025.5.10
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
